% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{beta_function}
\alias{beta_function}
\title{Beta density}
\usage{
beta_function(x, a, b)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{a}{First shape parameter.}

\item{b}{Second shape parameter.}
}
\value{
Density values.
}
\description{
Evaluate beta density at given points.
}
