% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_read_data.R
\name{change_class}
\alias{change_class}
\title{Change column class}
\usage{
change_class(
  dat = dat,
  factor_col = factor_col,
  number_col = setdiff(colnames(dat), factor_col)
)
}
\arguments{
\item{dat}{Input data frame.}

\item{factor_col}{Vector of factor column names.}

\item{number_col}{Vector of numeric column names.}
}
\value{
Data frame with updated classes.
}
\description{
Convert specified columns to factor or numeric.
}
