% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_read_data.R
\name{col1_to_rowname}
\alias{col1_to_rowname}
\title{First column to rownames}
\usage{
col1_to_rowname(data = data)
}
\arguments{
\item{data}{Input data frame.}
}
\value{
Data frame with updated rownames or error if duplicates.
}
\description{
Move first column to rownames and return data frame.
}
