% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function0_color.R
\name{darken_color}
\alias{darken_color}
\title{Darken or Lighten a Color}
\usage{
darken_color(color, darker = 0.1)
}
\arguments{
\item{color}{Input color.}

\item{darker}{Positive darkens, negative lightens.}
}
\value{
colors
}
\description{
Darken or Lighten a Color
}
