% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function0_color.R
\name{get_color}
\alias{get_color}
\title{Assign Colors to Categorical Levels}
\usage{
get_color(meta = meta, col = "celltype", pattern = "Set1")
}
\arguments{
\item{meta}{Metadata data frame.}

\item{col}{Column name to color.}

\item{pattern}{RColorBrewer palette name.}
}
\value{
colors
}
\description{
Assign Colors to Categorical Levels
}
