% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function0_color.R
\name{get_color_cluster_new}
\alias{get_color_cluster_new}
\title{Color Sub-clusters by Cell Type}
\usage{
get_color_cluster_new(meta = meta, celltype_color = celltype_color, r = 0.2)
}
\arguments{
\item{meta}{Metadata.}

\item{celltype_color}{Named color vector of celltypes.}

\item{r}{Blend ratio (0–1) between celltype color and rainbow.}
}
\value{
colors
}
\description{
Color Sub-clusters by Cell Type
}
