% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4_cluster.R
\name{get_hclust}
\alias{get_hclust}
\title{Hierarchical clustering}
\usage{
get_hclust(matrix = matrix, method = method, n_cluster = NULL, cut = NULL)
}
\arguments{
\item{matrix}{VAF matrix.}

\item{method}{Distance metric.}

\item{n_cluster}{Number of clusters.}

\item{cut}{Height cutoff.}
}
\value{
Vector of cluster labels.
}
\description{
Cluster cells by VAF distance and cut tree.
}
