% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4_cluster.R
\name{get_hclust_celltype}
\alias{get_hclust_celltype}
\title{Celltype-wise clustering}
\usage{
get_hclust_celltype(matrix = matrix, meta = meta)
}
\arguments{
\item{matrix}{VAF matrix.}

\item{meta}{Metadata with celltype.}
}
\value{
Metadata with appended cluster_new column.
}
\description{
Cluster cells within each celltype and append new cluster labels.
}
