% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{get_mtry}
\alias{get_mtry}
\title{Tune RF mtry}
\usage{
get_mtry(X_train = X_train, y_train = y_train, n_cv = 5)
}
\arguments{
\item{X_train}{Feature matrix.}

\item{y_train}{Response factor.}

\item{n_cv}{Number of folds.}
}
\value{
caret train object.
}
\description{
Cross-validate mtry for Random Forest.
}
