% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{get_ntree}
\alias{get_ntree}
\title{Tune RF ntree}
\usage{
get_ntree(
  X_train = X_train,
  y_train = y_train,
  n_cv = 5,
  mtry = mtry,
  ntree_list = seq(100, 1000, 100)
)
}
\arguments{
\item{X_train}{Feature matrix.}

\item{y_train}{Response factor.}

\item{n_cv}{Folds.}

\item{mtry}{Fixed mtry.}

\item{ntree_list}{Sequence of ntree values.}
}
\value{
List of caret models.
}
\description{
Cross-validate ntree list for fixed mtry.
}
