% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{get_ntree_best}
\alias{get_ntree_best}
\title{Select best ntree}
\usage{
get_ntree_best(cv_model_ntree = cv_model_ntree)
}
\arguments{
\item{cv_model_ntree}{List from get_ntree.}
}
\value{
Optimal ntree integer.
}
\description{
Pick ntree with max accuracy/kappa from CV results.
}
