% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{infer_AB}
\alias{infer_AB}
\title{Infer AB parameter}
\usage{
infer_AB(sum_alt, sum_ref, capture)
}
\arguments{
\item{sum_alt}{Alt read counts.}

\item{sum_ref}{Ref read counts.}

\item{capture}{Capture efficiency.}
}
\value{
AB value.
}
\description{
Estimate transformation parameter from alt/ref counts and capture rate.
}
