% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5_smooth.R
\name{infer_K}
\alias{infer_K}
\title{Compute K value}
\usage{
infer_K(var_string, ref = ref, sig = sig, K_smooth = 0.2)
}
\arguments{
\item{var_string}{Variant string "chr_pos_ref>alt".}

\item{ref}{Genome build hg19/hg38/mm10.}

\item{sig}{Signature 96-vector.}

\item{K_smooth}{Smoothing parameter.}
}
\value{
K value between 0 and 1.
}
\description{
Estimate signature-based K for a given variant.
}
