% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{infer_vaf}
\alias{infer_vaf}
\title{Infer true VAF}
\usage{
infer_vaf(ALT = ALT, REF = REF, AB = AB, lower = 0, upper = 1)
}
\arguments{
\item{ALT}{Alt read count.}

\item{REF}{Ref read count.}

\item{AB}{Transformation parameter.}

\item{lower}{Integration lower bound.}

\item{upper}{Integration upper bound.}
}
\value{
Estimated VAF.
}
\description{
Integrate transformed beta posterior to estimate allele fraction.
}
