% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4_cluster.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{Compute distance matrix}
\usage{
make_distance_matrix(matrix_VAF = matrix_VAF, method = method)
}
\arguments{
\item{matrix_VAF}{VAF matrix (variants × cells).}

\item{method}{Distance metric.}
}
\value{
Symmetric distance matrix.
}
\description{
Calculate cell-to-cell distance from VAF matrix.
}
