% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5_smooth.R
\name{make_smooth_matrix}
\alias{make_smooth_matrix}
\title{Smooth VAF matrix}
\usage{
make_smooth_matrix(
  matrix_VAF,
  matrix_REF,
  matrix_ALT,
  meta,
  leader_min_depth,
  neighbor_min_dis,
  ref,
  sig = NULL,
  method = method,
  K_smooth = 0.2
)
}
\arguments{
\item{matrix_VAF}{VAF matrix.}

\item{matrix_REF}{REF count matrix.}

\item{matrix_ALT}{ALT count matrix.}

\item{meta}{Metadata with celltype and cluster_tem.}

\item{leader_min_depth}{Min depth for leader cells.}

\item{neighbor_min_dis}{Max distance for neighbors.}

\item{ref}{Genome build.}

\item{sig}{Optional 96-signature vector.}

\item{method}{Distance metric.}

\item{K_smooth}{Smoothing parameter.}
}
\value{
Smoothed VAF matrix.
}
\description{
Smooth VAF values using local leaders/neighbors and signature weight.
}
