% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{make_VAF_matrix}
\alias{make_VAF_matrix}
\title{Build VAF matrix}
\usage{
make_VAF_matrix(matrix_REF, matrix_ALT, fast_cutoff = 0.9)
}
\arguments{
\item{matrix_REF}{REF count matrix.}

\item{matrix_ALT}{ALT count matrix.}

\item{fast_cutoff}{Proportion threshold for fast 0/1/2 encoding.}
}
\value{
VAF matrix.
}
\description{
Estimate VAF from REF/ALT count matrices via transformation.
}
