% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{mutation_filter}
\alias{mutation_filter}
\title{Filter mutations}
\usage{
mutation_filter(
  mut = mut,
  min_celltype = 10,
  min_mut_per_cell = 10,
  min_cell_one_mut = 10,
  meta = meta
)
}
\arguments{
\item{mut}{Mutation data frame.}

\item{min_celltype}{Min cells per cell type.}

\item{min_mut_per_cell}{Min mutations per cell.}

\item{min_cell_one_mut}{Min cells sharing one mutation.}

\item{meta}{Metadata data frame.}
}
\value{
Filtered mutation data frame.
}
\description{
Iteratively filter cells and mutations by thresholds.
}
