% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_read_data.R
\name{read_annovar}
\alias{read_annovar}
\title{Read ANNOVAR outputs}
\usage{
read_annovar(path = path, col_keep = col_keep, col_name = NULL)
}
\arguments{
\item{path}{Folder containing *.txt files.}

\item{col_keep}{Column indices to retain.}

\item{col_name}{Optional new column names vector.}
}
\value{
Merged data frame with cell column.
}
\description{
Batch read ANNOVAR txt files and merge into one table.
}
