% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4_cluster.R
\name{refine_meta}
\alias{refine_meta}
\title{Refine metadata}
\usage{
refine_meta(meta = meta, min_cell = min_cell)
}
\arguments{
\item{meta}{Metadata data frame.}

\item{min_cell}{Minimum cell count per group.}
}
\value{
Updated metadata with cluster_tem refined.
}
\description{
Remove low-count celltypes/clusters and re-label ambiguous clusters.
}
