% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4_cluster.R
\name{refine_meta_final}
\alias{refine_meta_final}
\title{Finalize cluster labels}
\usage{
refine_meta_final(meta = meta)
}
\arguments{
\item{meta}{Metadata data frame.}
}
\value{
Metadata with sequential cluster_new labels.
}
\description{
Re-label clusters within each celltype sequentially.
}
