% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{rounding_VAF}
\alias{rounding_VAF}
\title{Round VAF matrix}
\usage{
rounding_VAF(matrix)
}
\arguments{
\item{matrix}{VAF numeric matrix.}
}
\value{
Integer matrix.
}
\description{
Discretize VAF to 0/1/2 states.
}
