% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{show_imp}
\alias{show_imp}
\title{Plot feature importance}
\usage{
show_imp(dat = dat, top = 50)
}
\arguments{
\item{dat}{Data frame of importance scores.}

\item{top}{Number of top features.}
}
\value{
ggplot object.
}
\description{
ggplot of top features colored by mutation type.
}
