% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{show_mtry}
\alias{show_mtry}
\title{Plot mtry curve}
\usage{
show_mtry(dat = dat)
}
\arguments{
\item{dat}{Data frame from caret results.}
}
\value{
ggplot object.
}
\description{
ggplot of CV accuracy vs mtry.
}
