% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{show_ntree}
\alias{show_ntree}
\title{Plot ntree curve}
\usage{
show_ntree(dat = dat, celltype_color = celltype_color)
}
\arguments{
\item{dat}{Long-format data frame.}

\item{celltype_color}{Named color vector.}
}
\value{
ggplot object.
}
\description{
ggplot of CV error vs ntree by celltype.
}
