% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function6_randomforest.R
\name{show_prob}
\alias{show_prob}
\title{Plot margin probability}
\usage{
show_prob(dat = dat, celltype_color = celltype_color)
}
\arguments{
\item{dat}{Data frame with Prob and celltype.}

\item{celltype_color}{Named color vector.}
}
\value{
ggplot object.
}
\description{
ggplot of RF margin vs index.
}
