% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2_signature.R
\name{show_signature}
\alias{show_signature}
\title{Plot 96-trinucleotide signature}
\usage{
show_signature(array = array, name = name)
}
\arguments{
\item{array}{Named numeric vector of signature counts.}

\item{name}{Sample name for plot title.}
}
\value{
List of signature table and ggplot object.
}
\description{
Generate COSMIC-style barplot for 96 signatures and return table.
}
