% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2_signature.R
\name{snv_to_sig}
\alias{snv_to_sig}
\title{Convert SNV to signature}
\usage{
snv_to_sig(dat = dat, ref = ref)
}
\arguments{
\item{dat}{SNV data frame.}

\item{ref}{Genome build hg19/hg38/mm10.}
}
\value{
Data frame with context, alteration and signature columns.
}
\description{
Extract trinucleotide context and signature from SNV table.
}
