% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_read_data.R
\name{split_FORMAT_col}
\alias{split_FORMAT_col}
\title{Split FORMAT-like column}
\usage{
split_FORMAT_col(dat = dat, split = 2, name = 1, sep = ":")
}
\arguments{
\item{dat}{Input data frame.}

\item{split}{Column index of values to split.}

\item{name}{Column index of field names.}

\item{sep}{Separator character.}
}
\value{
Data frame with named numeric columns.
}
\description{
Split colon-separated FORMAT field into named numeric columns.
}
