% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_read_data.R
\name{split_INFO_col}
\alias{split_INFO_col}
\title{Split INFO-like column}
\usage{
split_INFO_col(dat = dat, split = split, sep = ";", link = "=")
}
\arguments{
\item{dat}{Input data frame.}

\item{split}{Column index to split.}

\item{sep}{Separator between fields.}

\item{link}{Separator between key and value.}
}
\value{
Data frame with named columns.
}
\description{
Split semicolon-separated INFO field into key=value columns.
}
