% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1_read_data.R
\name{split_VAF_col}
\alias{split_VAF_col}
\title{Split VAF column}
\usage{
split_VAF_col(dat = dat, col = col, into = NA, sep = ",")
}
\arguments{
\item{dat}{Input data frame.}

\item{col}{Column index to split.}

\item{into}{New column names vector or NA for auto naming.}

\item{sep}{Separator character.}
}
\value{
Data frame with split columns.
}
\description{
Split comma-separated VAF into multiple numeric columns.
}
