% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function0_statistic.R
\name{statistic_single_mut}
\alias{statistic_single_mut}
\title{Single-cell mutation statistics}
\usage{
statistic_single_mut(sta = sta, ref = ref, bar = "se")
}
\arguments{
\item{sta}{Mutation data frame.}

\item{ref}{Genome build hg19/hg38/mm10.}

\item{bar}{Error bar type se or sd.}
}
\value{
ggplot object
}
\description{
Plot mutation types, substitutions, RNA editing, dbSNP and cumulative variants.
}
