% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3_make_vaf.R
\name{transform_function}
\alias{transform_function}
\title{Transform allele fraction}
\usage{
transform_function(x, AB)
}
\arguments{
\item{x}{Numeric vector in (0,1).}

\item{AB}{Transformation parameter.}
}
\value{
Transformed values.
}
\description{
Apply nonlinear transformation to allele fraction.
}
